Start on prototype of a new paging implementation
The new paging implementation would make the default behavior for
callers hitting various paginated methods a single unpaginated call.
In order to get the paginated version, callers have to explicitly ask
for, e.g. `TransferClient.paginated.endpoint_search`.
In fact, `TransferClient.paginated` is a collection object containing
"shadow" implementations of the various paginated methods with
pagination enabled.

Paginated methods don't return a specialized resource object which
allows direct iteration over results. Instead, callers are expected to
walk the pages and then do a nested walk over the internal results.

The result is a much more bare-bones pagination implementation which
should be easier to port amongst APIs and which exposes the (sometimes
important) information which wraps a page of results. The current
PaginatedResource implementation is implicitly dropping this
information by walking over data in an internal key (in Transfer, this
is always `"DATA"`).